+++ CHANGELOG.md
@@ -4,10 +4,11 @@
 - Drop support for Windows 7 and 8
 - Drop support for macOS 10.13 and 10.14
 - Add playback history to queue panel
-- Add queue slide-out transition
+- Redesign queue panel
+- Add queue panel slide-out transition
 - Fix dragging tracks to up next/autoplay queue boundary
 - Fix button edge not clickable due to zoom
-- Update dependencies
+- Make queue dragline stable during scroll
 
 ## 0.18.0 - 2024 Jun 7
 - Add arm64 support for macOS and Windows